Extract `EventsPage` component and reuse it in `home` and `events` routes to improve code reusability.
